- adding money supply parameter to getblock rpc commands - refactoring tests a bit, made progress on blockchain tests, but still not working to satisfaction.
